+++ selene/elements.py
@@ -1,23 +1,27 @@
+import warnings
 from _ast import Tuple, List
 from collections import Sequence
 
-from selenium.common.exceptions import NoSuchElementException
+from future.utils import with_metaclass
+from selenium.common.exceptions import NoSuchElementException, TimeoutException
 from selenium.webdriver import ActionChains
 from selenium.webdriver.common.by import By
 from selenium.webdriver.common.keys import Keys
 
+import selene
+import selene.tools
 from selene import config
 from selene.abctypes.locators import ISeleneWebElementLocator, ISeleneListWebElementLocator
 from selene.abctypes.search_context import ISearchContext
 from selene.abctypes.webdriver import IWebDriver
 from selene.abctypes.webelement import IWebElement
-from selene.conditions import Condition, not_
 from selene.common.delegation import DelegatingMeta
 from selene.helpers import css_or_by_to_by
 from selene.support import by
 from selene.support.conditions import be
 from selene.support.conditions import have
 from selene.wait import wait_for
+from selene.conditions import not_, is_matched
 
 try:
     from functools import lru_cache
@@ -51,7 +55,8 @@ class InnerWebElementLocator(ISeleneWebElementLocator):
         return "%s.find_by%s" % (self._element, self._by)
 
     def find(self):
-        return self._element.get_actual_webelement().find_element(*self._by)
+        # return self._element.get_actual_webelement().find_element(*self._by)
+        return wait_for(self._element, be.visible, config.timeout, config.poll_during_waits).find_element(*self._by)
 
 
 class CachingWebElementLocator(ISeleneWebElementLocator):
@@ -59,9 +64,10 @@ class CachingWebElementLocator(ISeleneWebElementLocator):
     def description(self):
         return "Caching %s" % (self._element,)
 
+    # todo: will it cash kine of "first wrong webelement"? i.e. invisible element
     @lru_cache()
     def find(self):
-        return self._element()
+        return self._element.get_actual_webelement()
 
     def __init__(self, element):
         self._element = element
@@ -71,7 +77,12 @@ class CachingWebElementLocator(ISeleneWebElementLocator):
 # todo: Should we use this order convention? like below...
 class IndexedWebElementLocator(ISeleneWebElementLocator):
     def find(self):
-        return self._collection.get_actual_webelements()[self._index]
+        # return self._collection.get_actual_webelements()[self._index]
+        return wait_for(
+            self._collection,
+            have.size_at_least(self._index + 1),
+            config.timeout,
+            config.poll_during_waits)[self._index]
 
     @property
     def description(self):
@@ -108,7 +119,9 @@ class InnerListWebElementLocator(ISeleneListWebElementLocator):
         return "(%s).find_all_by(%s)" % (self._element, self._by)
 
     def find(self):
-        return self._element.get_actual_webelement().find_elements(*self._by)
+        # return self._element.get_actual_webelement().find_elements(*self._by)
+        return wait_for(self._element, be.visible, config.timeout, config.poll_during_waits)\
+            .find_elements(*self._by)
 
 
 class FilteredListWebElementLocator(ISeleneListWebElementLocator):
@@ -116,7 +129,7 @@ class FilteredListWebElementLocator(ISeleneListWebElementLocator):
         webelements = self._collection()
         filtered = [webelement
                     for webelement in webelements
-                    if self._condition.matches_webelement(webelement)]
+                    if is_matched(self._condition, webelement)]
         return filtered
 
     @property
@@ -131,8 +144,11 @@ class FilteredListWebElementLocator(ISeleneListWebElementLocator):
 
 class SlicedListWebElementLocator(ISeleneListWebElementLocator):
     def find(self):
-        # self._collection.should(have.size_at_least(self._slice.stop)) # todo: remove once covered with tests
-        webelements = self._collection()
+        # webelements = self._collection()
+        webelements = wait_for(self._collection,
+                               have.size_at_least(self._slice.stop),
+                               config.timeout,
+                               config.poll_during_waits)
         return webelements[self._slice.start:self._slice.stop:self._slice.step]
 
     @property
@@ -148,7 +164,7 @@ class SlicedListWebElementLocator(ISeleneListWebElementLocator):
 class FoundByConditionWebElementLocator(ISeleneWebElementLocator):
     def find(self):
         for webelement in self._collection():
-            if self._condition.matches_webelement(webelement):
+            if is_matched(self._condition, webelement):
                 return webelement
         raise NoSuchElementException('Element was not found by: %s' % (self._condition,))
 
@@ -162,8 +178,21 @@ class FoundByConditionWebElementLocator(ISeleneWebElementLocator):
         self._collection = collection
 
 
-class SeleneElement(IWebElement):
-    __metaclass__ = DelegatingMeta
+def _wait_with_screenshot(entity, condition, timeout=None, polling=None):
+    if timeout is None:
+        timeout = config.timeout
+    if polling is None:
+        polling = config.poll_during_waits
+    try:
+        return wait_for(entity, condition, timeout, polling)
+    except TimeoutException as e:
+        screenshot = selene.tools.take_screenshot()
+        msg = '''{original_msg}
+            screenshot: {screenshot}'''.format(original_msg=e.msg, screenshot=screenshot)
+        raise TimeoutException(msg, e.screen, e.stacktrace)
+
+
+class SeleneElement(with_metaclass(DelegatingMeta, IWebElement)):
 
     @property
     def __delegate__(self):
@@ -218,6 +247,9 @@ class SeleneElement(IWebElement):
     def __str__(self):
         return self._locator.description
 
+    def _execute_on_webelement(self, command, condition=be.or_not_to_be):
+        return command(_wait_with_screenshot(self, condition))
+
     # *** Relative elements ***
 
     def element(self, css_selector_or_by):
@@ -266,50 +298,60 @@ class SeleneElement(IWebElement):
     # *** Asserts (Explicit waits) ***
 
     def should(self, condition, timeout=None):
-        if not timeout:
+        if timeout is None:
             timeout = config.timeout
         # todo: implement proper cashing
-        # self._found = wait_for(self, condition, condition, timeout)
-        wait_for(self, condition, condition, timeout)
+        # self._found = wait_for(self, condition, timeout)
+        _wait_with_screenshot(self, condition, timeout)
         return self
 
     # todo: consider removing some aliases
-    insist = should
     assure = should
     should_be = should
     should_have = should
 
+    def insist(self, condition, timeout=None):
+        """ Deprecated use should or assure instead
+        """
+        warnings.warn("use should or assure instead", DeprecationWarning)
+        return self.should(condition, timeout)
+
     def should_not(self, condition, timeout=None):
-        if not timeout:
+        if timeout is None:
             timeout = config.timeout
         # todo: implement proper cashing
         not_condition = not_(condition)
-        wait_for(self, not_condition, not_condition, timeout)
+        _wait_with_screenshot(self, not_condition, timeout)
         return self
 
     # todo: consider removing some aliases
-    insist_not = should_not
     assure_not = should_not
     should_not_be = should_not
     should_not_have = should_not
 
+    def insist_not(self, condition, timeout=None):
+        """ Deprecated use should or assure instead
+        """
+        warnings.warn("use should_not or assure_not instead", DeprecationWarning)
+        return self.should_not(condition, timeout)
+
     # *** Additional actions ***
 
     def double_click(self):
-        self._execute(
-            lambda: self._actions_chains.double_click(self.__delegate__).perform(),
-            condition=be.Visible())
+        self._execute_on_webelement(
+            lambda it: self._actions_chains.double_click(it).perform(),
+            condition=be.visible)
         return self
 
     def set(self, new_text_value):
 
-        def clear_and_send_keys():
-            self.__delegate__.clear()
-            self.__delegate__.send_keys(new_text_value)
+        def clear_and_send_keys(webelement):
+            webelement.clear()
+            webelement.send_keys(new_text_value)
 
-        self._execute(
+        self._execute_on_webelement(
             clear_and_send_keys,
-            condition=be.Visible())
+            condition=be.visible)
 
         return self
 
@@ -325,160 +367,149 @@ class SeleneElement(IWebElement):
         return self.send_keys(Keys.TAB)
 
     def hover(self):
-        self._execute(
-            lambda: self._actions_chains.move_to_element(self.__delegate__).perform(),
-            condition=be.Visible())
+        self._execute_on_webelement(
+            lambda it: self._actions_chains.move_to_element(it).perform(),
+            condition=be.visible)
         return self
 
     # *** ISearchContext methods ***
 
     def find_elements(self, by=By.ID, value=None):
-        return self._execute(
-            lambda: self.__delegate__.find_elements(by, value),
-            condition=be.Visible())
+        return self._execute_on_webelement(
+            lambda it: it.find_elements(by, value),
+            condition=be.visible)
         # return self.__delegate__.find_elements(by, value) # todo: remove
 
     def find_element(self, by=By.ID, value=None):
-        return self._execute(
-            lambda: self.__delegate__.find_element(by, value),
-            condition=be.Visible())
+        return self._execute_on_webelement(
+            lambda it: it.find_element(by, value),
+            condition=be.visible)
         # return self.__delegate__.find_element(by, value) # todo: remove
 
     # *** IWebElement methods ***
 
     @property
     def tag_name(self):
-        return self._execute(
-            lambda: self.__delegate__.tag_name,
+        return self._execute_on_webelement(
+            lambda it: it.tag_name,
             condition=be.in_dom)
 
     @property
     def text(self):
-        return self._execute(
-            lambda: self.__delegate__.text,
-            condition=be.Visible())
+        return self._execute_on_webelement(
+            lambda it: it.text,
+            condition=be.visible)
 
     def click(self):
-        self._execute(
-            lambda: self.__delegate__.click(),
-            condition=be.Visible())
+        self._execute_on_webelement(
+            lambda it: it.click(),
+            condition=be.visible)
         return self  # todo: think on: IWebElement#click was supposed to return None
 
     def submit(self):
-        self._execute(
-            lambda: self.__delegate__.submit(),
-            condition=be.Visible())
+        self._execute_on_webelement(
+            lambda it: it.submit(),
+            condition=be.visible)
         return self
 
     def clear(self):
-        self._execute(
-            lambda: self.__delegate__.clear(),
-            condition=be.Visible())
+        self._execute_on_webelement(
+            lambda it: it.clear(),
+            condition=be.visible)
         return self
 
     def get_attribute(self, name):
-        return self._execute(
-            lambda: self.__delegate__.get_attribute(name),
+        return self._execute_on_webelement(
+            lambda it: it.get_attribute(name),
             condition=be.in_dom)
 
     def is_selected(self):
-        return self._execute(
-            lambda: self.__delegate__.is_selected(),
-            condition=be.Visible())
+        return self._execute_on_webelement(
+            lambda it: it.is_selected(),
+            condition=be.visible)
 
     def is_enabled(self):
-        return self._execute(
-            lambda: self.__delegate__.is_enabled(),
-            condition=be.Visible())
+        return self._execute_on_webelement(
+            lambda it: it.is_enabled(),
+            condition=be.visible)
 
     def send_keys(self, *value):
-        self._execute(
-            lambda: self.__delegate__.send_keys(*value),
-            condition=be.Visible())
+        self._execute_on_webelement(
+            lambda it: it.send_keys(*value),
+            condition=be.visible)
         return self
 
     # RenderedWebElement Items
     def is_displayed(self):
-        return self._execute(
-            lambda: self.__delegate__.is_displayed(),
+        return self._execute_on_webelement(
+            lambda it: it.is_displayed(),
             condition=be.in_dom)
 
     @property
     def location_once_scrolled_into_view(self):
-        return self._execute(
-            lambda: self.__delegate__.location_once_scrolled_into_view,
-            condition=be.Visible())
+        return self._execute_on_webelement(
+            lambda it: it.location_once_scrolled_into_view,
+            condition=be.visible)
 
     @property
     def size(self):
-        return self._execute(
-            lambda: self.__delegate__.size,
-            condition=be.Visible())
+        return self._execute_on_webelement(
+            lambda it: it.size,
+            condition=be.visible)
 
     def value_of_css_property(self, property_name):
-        return self._execute(
-            lambda: self.__delegate__.value_of_css_property(property_name),
+        return self._execute_on_webelement(
+            lambda it: it.value_of_css_property(property_name),
             condition=be.in_dom)
 
     @property
     def location(self):
-        return self._execute(
-            lambda: self.__delegate__.location,
-            condition=be.Visible())
+        return self._execute_on_webelement(
+            lambda it: it.location,
+            condition=be.visible)
 
     @property
     def rect(self):
-        return self._execute(
-            lambda: self.__delegate__.rect,
-            condition=be.Visible())
+        return self._execute_on_webelement(
+            lambda it: it.rect,
+            condition=be.visible)
 
     @property
     def screenshot_as_base64(self):
-        return self._execute(
-            lambda: self.__delegate__.screenshot_as_base64,
-            condition=be.Visible())  # todo: or `be.in_dom`?
+        return self._execute_on_webelement(
+            lambda it: it.screenshot_as_base64,
+            condition=be.visible)  # todo: or `be.in_dom`?
 
     @property
     def screenshot_as_png(self):
-        return self._execute(
-            lambda: self.__delegate__.screenshot_as_png,
-            condition=be.Visible())  # todo: or `be.in_dom`?
+        return self._execute_on_webelement(
+            lambda it: it.screenshot_as_png,
+            condition=be.visible)  # todo: or `be.in_dom`?
 
     def screenshot(self, filename):
-        return self._execute(
-            lambda: self.__delegate__.screenshot(filename),
-            condition=be.Visible())  # todo: or `be.in_dom`?
+        return self._execute_on_webelement(
+            lambda it: it.screenshot(filename),
+            condition=be.visible)  # todo: or `be.in_dom`?
 
     @property
     def parent(self):
-        return self._execute(
-            lambda: self.__delegate__.parent,  # todo: should not we return here some Selene entity as search_context?
+        return self._execute_on_webelement(
+            lambda it: it.parent,  # todo: should not we return here some Selene entity as search_context?
             condition=be.in_dom)
 
     @property
     def id(self):
-        return self._execute(
-            lambda: self.__delegate__.id,
+        return self._execute_on_webelement(
+            lambda it: it.id,
             condition=be.in_dom)
 
-    # *** private methods ***
 
-    def _execute(self, command, condition=be.or_not_to_be):
-        try:
-            return command()
-        except Exception:
-            self.should(condition)
-            return command()
-
-
-class SeleneCollection(Sequence):
+class SeleneCollection(with_metaclass(DelegatingMeta, Sequence)):
     """
     To fully match Selenium, SeleneCollection should extend collection.abc.MutableSequence.
     But that's the place where we should be more restrictive.
     It is actually the Selenium, who should use "Sequence" instead of "MutableSequence" (list)
     """
-
-    __metaclass__ = DelegatingMeta
 
     @property
     def __delegate__(self):
@@ -532,44 +563,58 @@ class SeleneCollection(Sequence):
     # * this may break DelegatingMeta logic (because we will have multiple inheritance...)
     # * this will Inheritance... Should not we at least use Composition here?
     def should(self, condition, timeout=None):
-        if not timeout:
+        if timeout is None:
             timeout = config.timeout
-        wait_for(self, condition, condition, timeout)
+        _wait_with_screenshot(self, condition, timeout)
         return self
 
     # todo: consider removing some aliases
-    insist = should
     assure = should
     should_be = should
     should_have = should
 
+    def insist(self, condition, timeout=None):
+        """ Deprecated use should or assure instead
+        """
+        warnings.warn("use should or assure instead", DeprecationWarning)
+        return self.should(condition, timeout)
+
     def should_not(self, condition, timeout=None):
-        if not timeout:
+        if timeout is None:
             timeout = config.timeout
         # todo: implement proper cashing
         not_condition = not_(condition)
-        wait_for(self, not_condition, not_condition, timeout)
+        _wait_with_screenshot(self, not_condition, timeout)
         return self
 
     # todo: consider removing some aliases are even all of them
-    insist_not = should_not
     assure_not = should_not
     should_not_be = should_not
     should_not_have = should_not
 
+    def insist_not(self, condition, timeout=None):
+        """ Deprecated use should_not or assure_not instead
+        """
+        warnings.warn("use should_not or assure_not instead", DeprecationWarning)
+        return self.should_not(condition, timeout)
+
     def should_each(self, condition, timeout=None):
-        if not timeout:
+        if timeout is None:
             timeout = config.timeout
 
         for selement in self:
             selement.should(condition, timeout)
 
+    assure_each = should_each
+
     def should_each_not(self, condition, timeout=None):
-        if not timeout:
+        if timeout is None:
             timeout = config.timeout
 
         for selement in self:
             selement.should_not(condition, timeout)
+
+    assure_each_not = should_each_not
 
     def filtered_by(self, condition):
         return SeleneCollection(FilteredListWebElementLocator(condition, self), self._webdriver)
@@ -577,15 +622,33 @@ class SeleneCollection(Sequence):
     ss = filtered_by
     all_by = filtered_by
     filtered = filtered_by
+
+    filter = filtered_by
     filter_by = filtered_by
-    filterBy = filtered_by
+
+    def filterBy(self, condition):
+        """ Deprecated use filtered_by or all_by instead
+        """
+        warnings.warn("use filtered_by or all_by instead", DeprecationWarning)
+        return self.filtered_by(condition)
 
     def element_by(self, condition):
         return SeleneElement(FoundByConditionWebElementLocator(condition, self), self._webdriver)
 
     s = element_by
     find_by = element_by
-    findBy = element_by
+
+    def findBy(self, condition):
+        """ Deprecated use element_by instead
+        """
+        warnings.warn("use element_by instead", DeprecationWarning)
+        return self.element_by(condition)
+
+    def find(self, condition):
+        """ Deprecated use element_by instead
+        """
+        warnings.warn("use element_by instead", DeprecationWarning)
+        return self.element_by(condition)
 
     # *** Sequence methods ***
 
@@ -600,8 +663,7 @@ class SeleneCollection(Sequence):
         # todo: optimise to the following:
         #   return self.waifFor(size_at_least(0)),
         # where waitFor will return the result of condition application, not self like should
-        self.should(have.size_at_least(0))
-        return len(self.__delegate__)
+        return len(_wait_with_screenshot(self, have.size_at_least(0)))
 
     # *** Overriden Sequence methods ***