+++ selene/wait.py
@@ -1,21 +1,21 @@
 import selene
 import selene.tools
 from selene import config
-from selene.conditions import not_
-
-__author__ = 'yashaka'
+# from selene.elements import SeleneElement, SeleneCollection
 
 import time
 from selenium.common.exceptions import TimeoutException
 
+from selene.abctypes.conditions import IEntityCondition
+from selene.exceptions import ConditionMismatchException
 
-def wait_for(entity, method, message='', timeout=None):
-    if not timeout:
-        timeout = config.timeout
+
+def wait_for(entity, condition, timeout=4, polling=0.1):
+    # type: (object, IEntityCondition, int) -> object
     end_time = time.time() + timeout
     while True:
         try:
-            return method(entity)
+            return condition.fn(entity)
         except Exception as reason:
             reason_message = getattr(reason, 'msg', getattr(reason, 'message', ''))
             reason_string = '{name}: {message}'.format(name=reason.__class__.__name__, message=reason_message)
@@ -23,30 +23,23 @@ def wait_for(entity, method, message='', timeout=None):
             stacktrace = getattr(reason, 'stacktrace', None)
 
             if time.time() > end_time:
-                screenshot = selene.tools.take_screenshot()
                 raise TimeoutException('''
             failed while waiting {timeout} seconds
             to assert {condition}
-            {message}
+            for {entity}
 
-            reason: {reason}
-            screenshot: {screenshot}'''.format(
+            reason: {reason}'''.format(
                     timeout=timeout,
-                    condition=method.description(),
-                    message=message,
-                    reason=reason_string,
-                    screenshot=screenshot), screen, stacktrace)
+                    condition=condition.description(),
+                    entity=entity,
+                    reason=reason_string), screen, stacktrace)
 
-            time.sleep(config.poll_during_waits)
+            time.sleep(polling)
 
 
-def has(entity, method):
+def satisfied(entity, condition):
     try:
-        value = method(entity)
+        value = condition(entity)
         return value if value is not None else False
     except Exception as exc:
         return False
-
-
-def has_not(entity, method):
-    return has(entity, not_(method))